+++ routes/controllers/AuthRoutesController.js
@@ -1,11 +1,15 @@
 const User = require('./../../models/User')
+const bcrypt = require('bcrypt')
 
 class AuthRoutesController {
     async register(req, res) {
         const { login, password } = req.body
 
+        const hashedPassword = bcrypt.hashSync(password, 7)
+
         const user = new User({
-            login, password
+            login: login, 
+            password: hashedPassword
         })
 
         const candidate = await User.findOne({login: login})
@@ -23,11 +27,17 @@ class AuthRoutesController {
     async login(req, res) {
         const { login, password } = req.body
 
-        const candidate = await User.findOne({login, password})
+        const candidate = await User.findOne({login})
 
         if (!candidate) { 
             return res.json({
                 message: 'There isn`t such account'
+            })
+        }
+
+        if (!bcrypt.compareSync(password, candidate?.password)) {
+            return res.json({
+                message: 'Incorrect login or password'
             })
         }