+++ routes/websocket.routes.js
@@ -1,46 +1,24 @@
 const User = require('./../models/User')
 const Message = require('./../models/Message')
+const WSController = require('./controllers/WSController')
+const res = require('express/lib/response')
 
 async function messageHandler(message, ws, wss) {
     message = JSON.parse(message)
 
+    const user = await User.findOne(message.user)
+    if (!user) return res.json({
+        event: 'error',
+        code: 401
+    })
+
     switch (message.event) {
         case 'message/connect': 
-            ws.login = message.user.login
-
-            const incoming = await Message.find({to: ws.login})
-            const outgoing = await Message.find({from: ws.login})
-
-            const messages = incoming.concat(outgoing)
-
-            ws.send(JSON.stringify({
-                event: 'message/connect',
-                messages: messages
-            }))
-
+            WSController.connect(message, ws, wss)
             break
 
         case 'message/send':
-            const sentMessage = new Message(message.message)
-
-            sentMessage.isSend = true
-
-            ws.send(JSON.stringify({
-                event: 'message/send',
-                message: sentMessage
-            }))
-
-            wss.clients.forEach(client => {
-                if (client.login === sentMessage.to) {
-                    client.send(JSON.stringify({
-                        event: 'message/send',
-                        message: sentMessage
-                    }))
-                }
-            }) 
-
-            await sentMessage.save()
-
+            WSController.send(message, ws, wss)
             break
 
         default: console.log('SWITCH DEFAULT CASE')